Enables auto-merge before approving PR
Auto merge can only be enabled *before* requirements are met. If the
only requirement is approval, and the PR is approved first, then
auto merge cannot be enabled.

This reverses the order, so auto-merge is enabled first, and then the
PR is approved. This way auto-merge works when the only requirement
is approval.